Allow to override the log level while logging exceptions
While logging information during an exception `except` block, instead of
overwriting the log level with `CRITICAL`, the log level used by the
caller will be set instead.

This allows the caller to use:

```python
try:
    raise ValueError()
except Exception as exc:
    logger.info(f"I was expecting this error: {exc}")
```

The log message will be logged as `INFO` level, whereas it was logged as
`CRITICAL` before.

If an `exception` attribute is still passed to the logger, as in
`logger.info("something", exception=exc)`, the log level will be `INFO`
as expected but the error will also still be correctly reported (in
Error Reporting + the bubble in Cloud Logging)